+++ question/type/wq/quizzes/lib/com/wiris/quizzes/api/ui/AnswerField.interface.php
@@ -1,0 +1,7 @@
+<?php
+
+interface com_wiris_quizzes_api_ui_AnswerField extends com_wiris_quizzes_api_ui_QuizzesField{
+	function getAnswerFieldComponent($listener);
+	function setEditorInitialParams($parameters);
+	function getFieldType();
+}